+++ src/java/ManagerQueries/popularStock.java
@@ -38,7 +38,7 @@ public class popularStock extends HttpServlet{
         }
         
         try{
-            int e_ssn = userBean.getSocialSecurityNumber();
+            long e_ssn = userBean.getSocialSecurityNumber();
             
             String query = "call mostPopularStocks("+e_ssn+");";
             ResultSet res = CapitaBay.ExecuteQuery(query);
